requirements.txt
Appium-Python-Client
Pytest